+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Auto-select the client in the `Send` view when selected in the left-side panel (if no selected client)
 - Color-coded entity types
 - Border Z-index for various UI controls
 - Pursuit + Step ID in debug filter